# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Đây là tài liệu Swagger cho hệ thống.",
        "title": "API DATN 2025",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Kin",
            "email": "trannguyentrungkien1006@gmail.com"
        },
        "license": {
            "name": "Apache 2.0"
        },
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/categories": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "API lấy ra tất cả danh mục đồ đạc",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "categories"
                ],
                "summary": "Get categories",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/categorydto.GetCategoryResponseWrapper"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/enums.AppError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/enums.AppError"
                        }
                    }
                }
            }
        },
        "/items": {
            "get": {
                "description": "API bao gồm cả lọc, phân trang và sắp xếp",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "items"
                ],
                "summary": "Get items",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "example": 1,
                        "description": "Current page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "example": 10,
                        "description": "Number record per page",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "name",
                        "description": "Sort column (vd: name)",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "ASC",
                        "description": "Sort type: ASC hoặc DESC",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Trường lọc (vd: email, full_name)",
                        "name": "searchBy",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Giá trị lọc (vd:abc@gmail.com, John Doe)",
                        "name": "searchValue",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/itemdto.GetItemResponseWrapper"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/enums.AppError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/enums.AppError"
                        }
                    }
                }
            },
            "post": {
                "description": "API thêm item mới",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "items"
                ],
                "summary": "Create new item",
                "parameters": [
                    {
                        "description": "Create item info",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/itemdto.CreateItemRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Created item successfully",
                        "schema": {
                            "$ref": "#/definitions/itemdto.CreateItemResponseWrapper"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/enums.AppError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/enums.AppError"
                        }
                    }
                }
            }
        },
        "/items/{itemID}": {
            "get": {
                "description": "API lấy thông tin item theo ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "items"
                ],
                "summary": "Get item by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID item",
                        "name": "itemID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/itemdto.GetItemByIDResponseWrapper"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/enums.AppError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/enums.AppError"
                        }
                    }
                }
            },
            "delete": {
                "description": "API xóa item theo ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "items"
                ],
                "summary": "Delete item",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID item",
                        "name": "itemID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deleted item successfully",
                        "schema": {
                            "$ref": "#/definitions/itemdto.DeleteItemResponseWrapper"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/enums.AppError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/enums.AppError"
                        }
                    }
                }
            },
            "patch": {
                "description": "API cập nhật item",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "items"
                ],
                "summary": "Update item",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID item",
                        "name": "itemID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update item info",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/itemdto.UpdateItemRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated item successfully",
                        "schema": {
                            "$ref": "#/definitions/itemdto.UpdateItemResponseWrapper"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/enums.AppError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/enums.AppError"
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "description": "Đăng nhập người dùng với email và mật khẩu mạnh",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Login",
                "parameters": [
                    {
                        "description": "Dữ liệu đăng nhập",
                        "name": "login",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/authdto.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/authdto.LoginResponseWrapper"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/enums.AppError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/enums.AppError"
                        }
                    }
                }
            }
        },
        "/posts": {
            "get": {
                "description": "API bao gồm cả lọc, phân trang và sắp xếp",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "posts"
                ],
                "summary": "Get posts",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "example": 1,
                        "description": "Current page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "example": 10,
                        "description": "Number record of page",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "authorName, title, createdAt",
                        "description": "Sort column",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "ASC, DESC",
                        "description": "Sort type",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "1, 2, 3",
                        "description": "Pending:1, Rejected:2, Approved:3",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "1, 2, 3, 4",
                        "description": "GiveAwayOldItem:1, FoundItem:2, SeekLoseItem:3, Other:4",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Trường lọc (vd: email, fullName)",
                        "name": "searchBy",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Giá trị lọc (vd:abc@gmail.com, John Doe)",
                        "name": "searchValue",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/postdto.GetAdminPostResponseWrapper"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/enums.AppError"
                        }
                    }
                }
            },
            "post": {
                "description": "API tạo mới một post và trả về thông tin post + user + JWT",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "posts"
                ],
                "summary": "Create a new post",
                "parameters": [
                    {
                        "description": "Post creation payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/postdto.CreatePostRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/postdto.CreatePostResponseWrapper"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/enums.AppError"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/enums.AppError"
                        }
                    }
                }
            }
        },
        "/posts/{postID}": {
            "get": {
                "description": "API lấy bài viết theo id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "posts"
                ],
                "summary": "Get detail post",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID post",
                        "name": "postID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/postdto.GetDetailPostResponseWrapper"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/enums.AppError"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/enums.AppError"
                        }
                    }
                }
            },
            "patch": {
                "description": "API cập nhật bài viết kết hợp với patch",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "posts"
                ],
                "summary": "Update posts",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID post",
                        "name": "postID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update post info",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/postdto.UpdatePostRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated post successfully",
                        "schema": {
                            "$ref": "#/definitions/postdto.UpdatePostResponseWrapper"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/enums.AppError"
                        }
                    }
                }
            }
        },
        "/users": {
            "get": {
                "description": "API bao gồm cả lọc, phân trang và sắp xếp",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Get users",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "example": 1,
                        "description": "Current page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "example": 10,
                        "description": "Number record of page",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "name",
                        "description": "Sort column (vd: fullName, email)",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "ASC",
                        "description": "Sort type: ASC hoặc DESC",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Trường lọc (vd: email, fullName)",
                        "name": "searchBy",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Giá trị lọc (vd:abc@gmail.com, John Doe)",
                        "name": "searchValue",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/userdto.GetUserResponseWrapper"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/enums.AppError"
                        }
                    }
                }
            },
            "post": {
                "description": "API thêm người dùng",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Create user",
                "parameters": [
                    {
                        "description": "Create user info",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/userdto.CreateUserRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created user successfully",
                        "schema": {
                            "$ref": "#/definitions/userdto.CreateUserResponseWrapper"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/enums.AppError"
                        }
                    }
                }
            }
        },
        "/users/{userID}": {
            "get": {
                "description": "API get user by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Get user by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID user",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/userdto.GetUserByIDResponseWrapper"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/enums.AppError"
                        }
                    }
                }
            },
            "delete": {
                "description": "API delete user by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Delete user",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID user",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/userdto.DeleteUserResponseWrapper"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/enums.AppError"
                        }
                    }
                }
            },
            "patch": {
                "description": "API cập nhật người dùng",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Update user",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID user",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update user info",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/userdto.UpdateUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated user successfully",
                        "schema": {
                            "$ref": "#/definitions/userdto.UpdateUserResponseWrapper"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/enums.AppError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "categorydto.GetCategoryResponseWrapper": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/categorydto.GetCategoryResponse"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "enums.AppError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "error": {
                    "type": "string",
                    "example": "INVALID_QUERY_PARAM"
                },
                "message": {
                    "type": "string",
                    "example": "Name is require"
                }
            }
        },
...
        "itemdto.CreateItemRequest": {
            "type": "object",
            "required": [
                "categoryID",
                "name"
            ],
            "properties": {
                "categoryID": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "itemdto.CreateItemResponseWrapper": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/itemdto.CreateItemResponse"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "itemdto.DeleteItemResponseWrapper": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "itemdto.GetItemByIDResponseWrapper": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/itemdto.GetItemByIDResponse"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "itemdto.GetItemResponseWrapper": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/itemdto.GetItemResponse"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}